add fee logics

+++ contracts/mocks/MockERC20.sol
@@ -3,22 +3,26 @@ pragma solidity ^0.8.14;
 
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
-contract MockERC20 is ERC20{
+contract MockERC20 is ERC20 {
     uint8 private d;
 
-    constructor(string memory name, string memory symbol, uint8 _decimals) ERC20(name, symbol){
+    constructor(
+        string memory name,
+        string memory symbol,
+        uint8 _decimals
+    ) ERC20(name, symbol) {
         d = _decimals;
     }
 
-    function mint(address account, uint amount) external {
+    function mint(address account, uint256 amount) external {
         _mint(account, amount);
     }
 
-    function burn(address account, uint amount) external {
+    function burn(address account, uint256 amount) external {
         _burn(account, amount);
     }
 
-    function decimals() public view override returns (uint8){
+    function decimals() public view override returns (uint8) {
         return d;
     }
 }
